make sure all connections are closed properly on client side
When the server closed the connection and all data are read, the socket
is in „CLOSE-WAIT“ status … but in this combination „connected“ returns
false, even if the socket is not really closed on the client.

This is not a real problem because normally „CLOSE_WAIT“ sockets are
reused for new connections, but it may end up in problems any may cause
problems. (I try to tackle down the effect that sometimes no more
connections are possible …)

The solution is to call „stop“ in any case when the Stream should be
closed because with this the connection is closed correctly in any
case. A double call is handed by „stop“ automatically so don’t worry
about this.