Checks and delay before reading header infos
- Added some checks that stream variable is not NULL before using it
(just to make sure)
- To avoid header-read errors that happened sometimes implement a check
for available data with a maximum timeout of 1000ms (configurable via
defines in header file). With my default settings we wait 20 ms and
then check again if at least 17 characters are available to read (which
should be a normal HTTP-header first line). This is done at max 50
times before the read is tried